avancement ex00 cpp06 converter check int et float

+++ cpp06/ex00/Converter.hpp
@@ -47,8 +47,8 @@ class Converter
 	int _type;
 
 	char _valueTypeChar;
-	int _valueTypeInt;
-	float _valueTypeFloat;
+	double _valueTypeInt;
+	double _valueTypeFloat;
 	double _valueTypeDouble;
 
 	static const int _typeChar = 1;
